fix tasks endpoints

+++ src/tasks/handlers/deleteTaskHandler.ts
@@ -8,14 +8,22 @@ export const deleteTaskHandler = async (
   req: Request,
   res: Response
 ) => {
-  const model = body.model;
+  const model = await TaskModel.findOne({ _id: body.model._id });
+  if (model === null) {
+    return res
+      .status(404)
+      .send({ message: 'Model with that id does not exist.' });
+  }
+  if (model.ownerId.toString() !== req.userId?.toString()) {
+    return res.status(403).send({ message: 'Not authorized.' });
+  }
 
   if (model.status === 'queued') {
     await QueueItemModel.deleteOne({ taskId: model!._id });
   }
 
   await ProjectModel.updateOne(
-    { _id: body.project._id },
+    { _id: model.projectId },
     {
       $pull: { models: model._id },
     }
